feat(UIKIT-1004,ui,DashboardLayout): Поправлены тип и тест
