+++ packages/components/src/MenuOrganization/MenuOrganizations.stubs.ts
@@ -1,0 +1,79 @@
+export const CURRENT_ORGANIZATION = {
+  name: 'АО Калуга-Астрал',
+  inn: '1231312312',
+  kpp: '1231231231',
+  id: '2',
+};
+
+export const ORGANIZATIONS = [
+  {
+    group: 'Подключенные',
+    data: [
+      {
+        name: 'АО Калуга-Астрал с очень длинным длинным длинным длинным названием',
+        inn: '170598611048',
+        kpp: '777401697',
+        id: '1',
+      },
+      {
+        name: 'АО Копыта и рога',
+        inn: '3605726192',
+        kpp: '777401697',
+        id: '2',
+      },
+      {
+        name: 'АО Копыта и рога',
+        inn: '170598611048',
+        kpp: '777401697',
+        id: '3',
+      },
+      {
+        name: 'АО Копыта и рога',
+        inn: '3605726192',
+        kpp: '1231231231',
+        id: '4',
+      },
+    ],
+  },
+  {
+    group: 'Неактивные',
+    data: [
+      {
+        name: 'АО Калуга-Астрал',
+        inn: '1231312312',
+        kpp: '1231231231',
+        id: '1',
+      },
+      {
+        name: 'АО Копыта и рога',
+        inn: '1231312312',
+        kpp: '1231231231',
+        id: '2',
+      },
+      {
+        name: 'АО Копыта и рога',
+        inn: '1231312312',
+        kpp: '1231231231',
+        id: '3',
+      },
+      {
+        name: 'АО Копыта и рога',
+        inn: '1231312312',
+        kpp: '1231231231',
+        id: '4',
+      },
+      {
+        name: 'АО Копыта и рога',
+        inn: '1231312312',
+        kpp: '1231231231',
+        id: '5',
+      },
+      {
+        name: 'АО Копыта и рога',
+        inn: '1231312312',
+        kpp: '1231231231',
+        id: '6',
+      },
+    ],
+  },
+];
